filament-recognition-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
dashscope>=1.14.0
pillow>=10.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
